Moved solution file to root.
